update verify level with new profile
