--- server/src/routes/admin/middlewares/__mocks__/verify-admin-level.js
@@ -1,3 +1,0 @@
-export function verifyAdminLevel (req, res, next) {
-  return next()
-}
