+++ server/src/routes/admin/middlewares/__mocks__/verify-repartiteur-level.js
@@ -1,0 +1,3 @@
+export function verifyRepartiteurLevel (req, res, next) {
+  return next()
+}
