+++ server/src/routes/admin/middlewares/index.js
@@ -1,4 +1,4 @@
+export * from './verify-repartiteur-level'
+export * from './verify-repartiteur-departement'
 export * from './verify-admin-level'
-export * from './verify-admin-departement'
-export * from './verify-tech-level'
 export * from './verify-access-aurige'
