+++ server/src/routes/admin/middlewares/verify-access-aurige.js
@@ -16,7 +16,7 @@ export function verifyAccessAurige (req, res, next) {
   try {
     if (
       requestedAction === 'aurige' &&
-      userLevel < config.userStatusLevels.tech
+      userLevel < config.userStatusLevels.admin
     ) {
       throw new Error('Accès interdit')
     }
