+++ server/src/routes/admin/middlewares/verify-admin-level.spec.js
@@ -3,9 +3,9 @@ import express from 'express'
 import jwt from 'jsonwebtoken'
 
 import { verifyToken } from '../../middlewares'
-import { verifyAdminLevel } from './verify-admin-level'
 import { apiPrefix } from '../../../app'
 import config from '../../../config'
+import { verifyAdminLevel } from './verify-admin-level'
 
 const basicData = {
   email: 'user@example.com',
@@ -25,6 +25,14 @@ const adminToken = jwt.sign(adminData, config.secret, {
   expiresIn: '30s',
 })
 
+const techData = {
+  email: 'tech@example.com',
+  level: 2,
+}
+const techToken = jwt.sign(techData, config.secret, {
+  expiresIn: '30s',
+})
+
 const app = express()
 app.use(verifyToken, verifyAdminLevel)
 app.get(apiPrefix, (req, res) => res.json({ ok: true }))
@@ -34,7 +42,7 @@ describe('Verify-token', () => {
     await app.close()
   })
 
-  it('Should respond a 401', async () => {
+  it('Should respond a 401 for basic ', async () => {
     // When
     const { body, status } = await request(app)
       .get(apiPrefix)
@@ -47,12 +55,25 @@ describe('Verify-token', () => {
     expect(body).toHaveProperty('message', 'Accès interdit')
   })
 
-  it('Should respond a 200', async () => {
+  it('Should respond a 401  for admin ', async () => {
     // When
     const { body, status } = await request(app)
       .get(apiPrefix)
       .set('Accept', 'application/json')
       .set('Authorization', `Bearer ${adminToken}`)
+
+    // Then
+    expect(status).toBe(401)
+    expect(body).toHaveProperty('success', false)
+    expect(body).toHaveProperty('message', 'Accès interdit')
+  })
+
+  it('Should respond a 200', async () => {
+    // When
+    const { body, status } = await request(app)
+      .get(apiPrefix)
+      .set('Accept', 'application/json')
+      .set('Authorization', `Bearer ${techToken}`)
 
     // Then
     expect(status).toBe(200)
