+++ server/src/routes/admin/middlewares/verify-repartiteur-level.js
@@ -1,15 +1,15 @@
 import config from '../../../config'
 import { appLogger } from '../../../util'
 
-export function verifyTechLevel (req, res, next) {
+export function verifyRepartiteurLevel (req, res, next) {
   const userLevel = req.userLevel
   const loggerInfo = {
-    section: 'admin-token',
+    section: 'admin-repartiteur-token',
     action: 'check-level',
     user: req.userId,
   }
   try {
-    if (userLevel >= config.userStatusLevels.tech) {
+    if (userLevel >= config.userStatusLevels.repartiteur) {
       return next()
     }
     throw new Error('Accès interdit')
