+++ server/src/routes/admin/middlewares/verify-repartiteur-level.spec.js
@@ -3,9 +3,9 @@ import express from 'express'
 import jwt from 'jsonwebtoken'
 
 import { verifyToken } from '../../middlewares'
+import { verifyRepartiteurLevel } from './verify-repartiteur-level'
 import { apiPrefix } from '../../../app'
 import config from '../../../config'
-import { verifyTechLevel } from './verify-tech-level'
 
 const basicData = {
   email: 'user@example.com',
@@ -25,16 +25,8 @@ const adminToken = jwt.sign(adminData, config.secret, {
   expiresIn: '30s',
 })
 
-const techData = {
-  email: 'tech@example.com',
-  level: 2,
-}
-const techToken = jwt.sign(techData, config.secret, {
-  expiresIn: '30s',
-})
-
 const app = express()
-app.use(verifyToken, verifyTechLevel)
+app.use(verifyToken, verifyRepartiteurLevel)
 app.get(apiPrefix, (req, res) => res.json({ ok: true }))
 
 describe('Verify-token', () => {
@@ -42,7 +34,7 @@ describe('Verify-token', () => {
     await app.close()
   })
 
-  it('Should respond a 401 for basic ', async () => {
+  it('Should respond a 401', async () => {
     // When
     const { body, status } = await request(app)
       .get(apiPrefix)
@@ -55,25 +47,12 @@ describe('Verify-token', () => {
     expect(body).toHaveProperty('message', 'Accès interdit')
   })
 
-  it('Should respond a 401  for admin ', async () => {
-    // When
-    const { body, status } = await request(app)
-      .get(apiPrefix)
-      .set('Accept', 'application/json')
-      .set('Authorization', `Bearer ${adminToken}`)
-
-    // Then
-    expect(status).toBe(401)
-    expect(body).toHaveProperty('success', false)
-    expect(body).toHaveProperty('message', 'Accès interdit')
-  })
-
   it('Should respond a 200', async () => {
     // When
     const { body, status } = await request(app)
       .get(apiPrefix)
       .set('Accept', 'application/json')
-      .set('Authorization', `Bearer ${techToken}`)
+      .set('Authorization', `Bearer ${adminToken}`)
 
     // Then
     expect(status).toBe(200)
